Split up the token instructions
